Return an array with a DEFAULT entry.
This is a breaking change. Old code must be updated.

Previous versions of MBS2 were initialized with the following code in
the "bam.lua" file:

  local atEnv = {}
  atEnv.DEFAULT = require "mbs"()

With this commit the "mbs" module returns an array with an initialized
"DEFAULT" entry. In a nutshell the old initialization code is moved into
the "mbs" module. This reduces the code in "bam.lua" to the following:

  local atEnv = require 'mbs'